Added features for handling current date&time and session timer.
Added timer increment functionality. Once session class has properties/member variables, need to update so we can store multiple date&time and have timers for each session that is running. For now its just one until a session class is set up or if we want to do it all in MainWindow, can change to a array/vector later.